--- RecipeBook.ApiGateway.Api/Features/UserAccount/Models/AuthModel.cs
@@ -1,8 +1,0 @@
-﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccount.Models
-{
-    public class AuthModel
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
