--- RecipeBook.ApiGateway.Api/Features/UserAccount/Models/ExistingUserAccountModel.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace RecipeBook.ApiGateway.Api.Features.UserAccount.Models
-{
-    public class ExistingUserAccountModel : UserAccountModel
-    {
-        public Guid Id { get; set; }
-    }
-}
