--- RecipeBook.ApiGateway.Api/Features/UserAccount/Models/NewUserAccountModel.cs
@@ -1,6 +1,0 @@
-﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccount.Models
-{
-    public class NewUserAccountModel : UserAccountModel
-    {
-    }
-}
