--- RecipeBook.ApiGateway.Api/Features/UserAccount/Models/UserAccountModel.cs
@@ -1,9 +1,0 @@
-﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccount.Models
-{
-    public class UserAccountModel
-    {
-        public string Firstname { get; set; }
-        public string Lastname { get; set; }
-        public string Role { get; set; }
-    }
-}
