+++ RecipeBook.ApiGateway.Api/Features/UserAccounts/Mapping/UserAccountModelProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using RecipeBook.ApiGateway.Api.Features.UserAccounts.Models;
+using RecipeBook.CoreApp.Api.Features.UserAccounts.Models;
+
+namespace RecipeBook.ApiGateway.Api.Features.UserAccounts.Mapping
+{
+    public class UserAccountModelProfile : Profile
+    {
+        public UserAccountModelProfile()
+        {
+            CreateMap<AuthModel, AuthDto>();
+            CreateMap<UserAccountDto, ExistingUserAccountModel>();
+        }
+    }
+}
