+++ RecipeBook.ApiGateway.Api/Features/UserAccounts/Models/AuthModel.cs
@@ -1,0 +1,8 @@
+﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccounts.Models
+{
+    public class AuthModel
+    {
+        public string Username { get; set; }
+        public string Password { get; set; }
+    }
+}
