+++ RecipeBook.ApiGateway.Api/Features/UserAccounts/Models/ExistingUserAccountModel.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace RecipeBook.ApiGateway.Api.Features.UserAccounts.Models
+{
+    public class ExistingUserAccountModel : UserAccountModel
+    {
+        public Guid Id { get; set; }
+    }
+}
