+++ RecipeBook.ApiGateway.Api/Features/UserAccounts/Models/NewUserAccountModel.cs
@@ -1,0 +1,6 @@
+﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccounts.Models
+{
+    public class NewUserAccountModel : UserAccountModel
+    {
+    }
+}
