+++ RecipeBook.ApiGateway.Api/Features/UserAccounts/Models/UserAccountModel.cs
@@ -1,0 +1,9 @@
+﻿namespace RecipeBook.ApiGateway.Api.Features.UserAccounts.Models
+{
+    public class UserAccountModel
+    {
+        public string Firstname { get; set; }
+        public string Lastname { get; set; }
+        public string Role { get; set; }
+    }
+}
