--- RecipeBook.CoreApp.Api/Features/UserAccount/Models/AuthDto.cs
@@ -1,8 +1,0 @@
-﻿namespace RecipeBook.CoreApp.Api.Features.UserAccount.Models
-{
-    public class AuthDto
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
