+++ RecipeBook.CoreApp.Api/Features/UserAccounts/Models/AuthDto.cs
@@ -1,0 +1,8 @@
+﻿namespace RecipeBook.CoreApp.Api.Features.UserAccounts.Models
+{
+    public class AuthDto
+    {
+        public string Username { get; set; }
+        public string Password { get; set; }
+    }
+}
